Widen int16 to int32 for weights, activation, and eligibility
- Connection.Weight: int16 → int32
- Connection.Eligibility: int16 → int32
- Neuron.Activation/Baseline/Threshold: int16 → int32
- ClampAdd: int32 args with int64 intermediate
- All learning rules, plasticity, benchmarks, celegans updated
- Predictive PredictionScale: default 15 (was 15 for int16, same
  value works for int32 with typical weight magnitudes)
- Decay uses int64 intermediate for distance calculation
- Decay cap increased from 32 to 64 ticks (int32 range is wider)
- 89 tests passing, zero regressions
- XOR benchmark unchanged (plateau is dynamics, not precision)

+++ benchmark/benchmark.go
@@ -159,7 +159,7 @@ func (t *Tracker) PrintReport(w io.Writer, taskName, ruleName string) {
 
 // WeightStats computes mean and standard deviation for a slice of
 // int16 weights.
-func WeightStats(weights []int16) (mean, stddev float64) {
+func WeightStats(weights []int32) (mean, stddev float64) {
 	if len(weights) == 0 {
 		return 0, 0
 	}
